requirements.txt
alembic==1.2.1
Cerberus==1.3.2
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-JWT==0.3.2
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
optionaldict==0.1.1
pycparser==2.19
PyJWT==1.4.2
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
redis==3.3.11
requests==2.22.0
rq==1.3.0
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.6
wechatpy==1.8.3
Werkzeug==0.16.0
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
